Add a few schemas
