+++ backend_sanity/.eslintrc
@@ -1,3 +1,3 @@
 {
-  "extends": "@sanity/eslint-config-studio"
+  "extends": ["next/babel"]
 }
