+++ backend_sanity/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "sanity-project",
+  "name": "portfolio",
   "private": true,
   "version": "1.0.0",
   "main": "package.json",
@@ -23,6 +23,8 @@
     "styled-components": "^5.2.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.20.12",
+    "@babel/preset-env": "^7.20.2",
     "@sanity/eslint-config-studio": "^2.0.1",
     "eslint": "^8.6.0",
     "prettier": "^2.8.2",
