--- backend_sanity/sanity.cli.ts
@@ -1,8 +1,0 @@
-import {defineCliConfig} from 'sanity/cli'
-
-export default defineCliConfig({
-  api: {
-    projectId: 'i83e4ijq',
-    dataset: 'production'
-  }
-})
