+++ backend_sanity/schemas/index.js
@@ -1,0 +1,4 @@
+export const schemaTypes = []
+// import testimonial from "./testimonial"
+
+
