--- backend_sanity/schemas/index.ts
@@ -1,1 +1,0 @@
-export const schemaTypes = []
