+++ backend_sanity/schemas/testimonial.js
@@ -1,0 +1,30 @@
+export default {
+  name: 'testimonials',
+  title: 'Testimonials',
+  type: 'document',
+  fields: [
+    {
+      name: 'name',
+      title: 'Name',
+      type: 'string',
+    },
+    {
+      name: 'company',
+      title: 'Company',
+      type: 'string',
+    },
+    {
+      name: 'imageurl',
+      title: 'ImgURL',
+      type: 'image',
+      options: {
+        hotspot: true,
+      }
+    },
+    {
+      name: 'feedback',
+      title: 'Feedback',
+      type: 'string',
+    },
+  ],
+}
