--- portfolio/.eslintrc
@@ -1,3 +1,0 @@
-{
-  "extends": ["next/babel"]
-}
