--- portfolio/.gitignore
@@ -1,26 +1,0 @@
-# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
-
-# Dependencies
-/node_modules
-/.pnp
-.pnp.js
-
-# Compiled Sanity Studio
-/dist
-
-# Temporary Sanity runtime, generated by the CLI on every dev server start
-/.sanity
-
-# Logs
-/logs
-*.log
-
-# Coverage directory used by testing tools
-/coverage
-
-# Misc
-.DS_Store
-*.pem
-
-# Typescript
-*.tsbuildinfo
