--- portfolio/sanity.config.js
@@ -1,18 +1,0 @@
-import {defineConfig} from 'sanity'
-import {deskTool} from 'sanity/desk'
-import {visionTool} from '@sanity/vision'
-import {schemaTypes} from './schemas'
-
-export default defineConfig({
-  name: 'default',
-  title: 'Portfolio',
-
-  projectId: 'lggxlpb3',
-  dataset: 'portfolio',
-
-  plugins: [deskTool(), visionTool()],
-
-  schema: {
-    types: schemaTypes,
-  },
-})
