--- portfolio/schemas/index.js
@@ -1,1 +1,0 @@
-export const schemaTypes = []
