--- portfolio/schemas/testimonial.js
@@ -1,3 +1,0 @@
-export default {
-  
-}
